- working skeleton app

+++ src/main/java/com/linkacademy/commerceapp/domain/entity/Product.java
@@ -4,10 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Table
 @Entity
@@ -23,5 +20,5 @@ public class Product extends EntityWithUUID {
     private Type type;
     @Enumerated(EnumType.STRING)
     private Brand brand;
-    private int quantityInStock;
+    private long quantityInStock;
 }
